Big changes, permissions + authentication, generating jwt + api auth using the jwt
