+++ app/controllers/concerns/exception_handler.rb
@@ -3,11 +3,29 @@
 module ExceptionHandler
   extend ActiveSupport::Concern
 
+  class AuthenticationError < StandardError; end
   class InvalidToken < StandardError; end
 
-  rescue_from ExceptionHandler::InvalidToken, with: four_twenty_two
+  included do
+    rescue_from ExceptionHandler::InvalidToken, with: :four_twenty_two
+    rescue_from ExceptionHandler::AuthenticationError, with: :unauthorized_request
+    rescue_from CanCan::AccessDenied, with: :forbidden_request
+  end
 
+  private
+
+  # JSON response with message; Status code 422 - unprocessable entity
   def four_twenty_two(e)
     render json: { message: e.message }, status: :unprocessable_entity
+  end
+
+  # JSON response with message; Status code 401 - Unauthorized
+  def unauthorized_request(e)
+    render json: { message: e.message }, status: :unauthorized
+  end
+
+  # JSON response with message; Status code 403 - Unauthorized
+  def forbidden_request(e)
+    render json: { message: "Forbidden, #{e.message}" }, status: :forbidden
   end
 end
