+++ app/models/ability.rb
@@ -4,6 +4,13 @@ class Ability
   include CanCan::Ability
 
   def initialize(user)
+    can :manage, Address do |address|
+      address.user == user
+    end
+    can :create, User if user.nil?
+    can %i[read update destroy], User do |usr|
+      usr == user
+    end
     # Define abilities for the passed in user here. For example:
     #
     #   user ||= User.new # guest user (not logged in)
