+++ lib/json_web_token.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class JsonWebToken
-  HMAC_SECRET = Rails.application.secrets.secret_key_base
+  HMAC_SECRET = Rails.application.credentials.secret_key_base
 
   def self.encode(payload, expiry = 1.day.from_now)
     payload[:exp] = expiry.to_i
@@ -12,6 +12,6 @@ class JsonWebToken
     payload = JWT.decode(token, HMAC_SECRET)[0]
     HashWithIndifferentAccess.new payload
   rescue JWT::DecodeError => e
-    raise ExceptionHandler::InvalidToken, e.message
+    raise ExceptionHandler::InvalidToken, "invalid token #{e.message}"
   end
 end
